refactor(sdk): decompose PasswayClient into services
