+++ sdk/src/services/DataTransform/index.ts
@@ -1,0 +1,10 @@
+export class DataTransformService {
+  stringToUintArray = (str: string) => {
+    const textEncoder = new TextEncoder()
+    const uint8Array = textEncoder.encode(str)
+
+    return uint8Array
+  }
+}
+
+export const dataTransform = new DataTransformService()
