+++ sdk/src/types.ts
@@ -1,0 +1,5 @@
+export interface RegistrationConfig {
+  appName: string
+  userName: string
+  userDisplayName: string
+}
